Assets/StreamingAssets/SERVER/requirements.txt
absl-py==2.3.1
asgiref==3.8.1
astunparse==1.6.3
audioop-lts==0.2.1
audioread==3.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Django==5.2.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.0
namex==0.1.0
numba==0.61.2
numpy==2.2.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.32.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uv==0.7.12
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
